requirements.txt
boto3==1.24.71
python-dotenv==0.21.0
